solve-image-APIs/requirements.txt
azure-functions==1.0.0b3
azure-functions-worker==1.0.0b3
azure-storage-blob
grpcio==1.14.2
grpcio-tools==1.14.2
protobuf==3.6.1
pillow
six==1.12.0